# fastrand: optimize Uint32N()