```
name                old time/op  new time/op  delta
FastRandBuf-16      32.5ns ± 4%  18.8ns ± 7%  -42.02%  (p=0.000 n=9+10)
FastRandUint32N-16  0.91ns ± 1%  0.19ns ±12%  -78.83%  (p=0.000 n=9+10)
```

### What problem does this PR solve?

No.

### What is changed and how it works?

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```

Issue Number: ref #33240